Changes made to ensure CLR compliance Redundant code and variables removed
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS13]$/xBIM/XbimFramework/Dev/ModelServer;C34977

+++ Xbim.IO/XbimSurfaceStyle.cs
@@ -48,7 +48,7 @@ namespace Xbim.IO
 
         public IfcSurfaceStyle IfcSurfaceStyle(XbimModel model)
         {
-            if (IsIfcSurfaceStyle) return (IfcSurfaceStyle)model.Instances[(uint)styleId]; else return null;
+            if (IsIfcSurfaceStyle) return (IfcSurfaceStyle)model.Instances[styleId]; else return null;
         }
 
         public IfcType IfcType
